Crate parachains_runtimes_test_utils

Modules

test_cases

Module contains predefined test-case scenarios for Runtime with common functionality.

Structs

CollatorSessionKey

CollatorSessionKeys

ExtBuilder

Basic builder based on balances, collators and pallet_session.

RuntimeHelper

SlotDurations

Enums

GovernanceOrigin

Enum representing governance origin/location.

XcmReceivedFrom

Traits

BasicParachainRuntime

A set of traits for a minimal parachain runtime, that may be used in conjunction with the ExtBuilder and the RuntimeHelper.

Functions

assert_metadata

assert_total

mock_open_hrmp_channel

Helper function which emulates opening HRMP channel which is needed for XcmpQueue to pass.

Type Aliases

AccountIdOf

BalanceOf

RuntimeCallOf

RuntimeOriginOf

SessionKeysOf

ValidatorIdOf